First try to use pkg extension
